added AuthModel specific folder and created the interfaces for auth user and jwttoken
